requirements.txt
asgiref==3.11.0
Django==5.2.9
django-ckeditor-5==0.2.18
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.19
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
django-allauth==65.3.1
requests==2.32.3
pyjwt==2.10.1
cryptography==44.0.0
whitenoise==6.8.2
pytest==8.3.4
pytest-django==4.9.0
